+++ src/icemac/ab/calendar/tests/test_recurrence.py
@@ -5,12 +5,6 @@ import unittest
 class RecurrenceMixIn(object):
     """Mix-in to test ..recurrence.*."""
 
-    def setUp(self):
-        super(RecurrenceMixIn, self).setUp()
-        self.recurrence_start = self.get_datetime((2013, 5, 3, 21, 45))
-        self.interval_start = self.get_datetime((2014, 4, 1))
-        self.interval_end = self.get_datetime((2014, 4, 30))
-
     def callFUT(self, adapter_name, datetime=None, start=None, end=None):
         from icemac.ab.calendar.recurrence import get_recurrences
         datetime = self.recurrence_start if datetime is None else datetime
@@ -21,6 +15,12 @@ class RecurrenceMixIn(object):
 
 class WeeklyTests(RecurrenceMixIn, icemac.ab.calendar.testing.ZCMLTestCase):
     """Testing ..recurrence.Weekly"""
+
+    def setUp(self):
+        super(RecurrenceMixIn, self).setUp()
+        self.recurrence_start = self.get_datetime((2013, 5, 3, 21, 45))
+        self.interval_start = self.get_datetime((2014, 4, 1))
+        self.interval_end = self.get_datetime((2014, 4, 30))
 
     def test_instance_fulfills_IRecurringDateTime_interface(self):
         from zope.interface.verify import verifyObject
@@ -65,13 +65,83 @@ class WeeklyTests(RecurrenceMixIn, icemac.ab.calendar.testing.ZCMLTestCase):
                 end=self.get_datetime((2014, 4, 5))))
 
 
+class MonthlyNthWeekdayTests(RecurrenceMixIn,
+                             icemac.ab.calendar.testing.ZCMLTestCase):
+    """Testing ..recurrence.MonthlyNthWeekday"""
+
+    def setUp(self):
+        super(RecurrenceMixIn, self).setUp()
+        # 3rd Thuesday in month
+        self.recurrence_start = self.get_datetime((2013, 3, 21, 21, 45))
+        self.interval_start = self.get_datetime((2014, 4, 1, 0))
+        self.interval_end = self.get_datetime((2014, 4, 30, 0))
+
+    def test_instance_fulfills_IRecurringDateTime_interface(self):
+        from zope.interface.verify import verifyObject
+        from ..interfaces import IRecurringDateTime
+        from ..recurrence import MonthlyNthWeekday
+        self.assertTrue(
+            verifyObject(IRecurringDateTime, MonthlyNthWeekday(None)))
+
+    def test_returns_empty_interval_if_datetime_after_interval_end(self):
+        self.assertEqual(
+            [], self.callFUT(
+                'nth weekday of month',
+                datetime=self.get_datetime((2014, 5, 1, 21, 45))))
+
+    def test_returns_all_first_of_month_in_interval_for_same_weekday(self):
+        self.assertEqual(
+            [self.get_datetime((2014, 4, 17, 21, 45)),
+             self.get_datetime((2014, 5, 15, 21, 45)),
+             self.get_datetime((2014, 6, 19, 21, 45))],
+            self.callFUT('nth weekday of month',
+                         end=self.get_datetime((2014, 6, 30, 17))))
+        self.assertEqual(self.recurrence_start.isoweekday(),
+                         self.callFUT('nth weekday of month')[0].isoweekday())
+
+    def test_does_not_start_before_datetime(self):
+        self.assertEqual(
+            [self.get_datetime((2014, 5, 4, 21, 45)),
+             self.get_datetime((2014, 6, 1, 21, 45))],
+            self.callFUT('nth weekday of month',
+                         datetime=self.get_datetime((2014, 5, 4, 21, 45)),
+                         end=self.get_datetime((2014, 6, 30, 17))))
+        self.assertEqual(self.recurrence_start.isoweekday(),
+                         self.callFUT('nth weekday of month')[0].isoweekday())
+
+    def test_interval_end_does_not_belong_to_interval(self):
+        self.assertEqual(
+            [], self.callFUT('nth weekday of month',
+                             datetime=self.get_datetime((2014, 4, 30, 0))))
+
+    def test_interval_start_belongs_to_interval(self):
+        self.assertEqual(
+            [self.get_datetime((2014, 4, 1, 0))],
+            self.callFUT(
+                'nth weekday of month',
+                datetime=self.get_datetime((2014, 4, 1, 0))))
+
+    def test_does_not_swap_into_next_month_if_month_do_not_have_a_fifth_week(
+            self):
+        self.assertEqual(
+            [self.get_datetime((2014, 5, 31, 0)),
+             self.get_datetime((2014, 8, 30, 0))],
+            self.callFUT(
+                'nth weekday of month',
+                datetime=self.get_datetime((2014, 5, 31, 0)),
+                start=self.get_datetime((2014, 5, 1, 0)),
+                end=self.get_datetime((2014, 8, 31, 0))))
+
+    # Test: 5. Sonntag im Monat faellt manchmal aus!
+
+
 class NextDateOfSameWeekdayTests(unittest.TestCase,
                                  icemac.ab.calendar.testing.TestMixIn):
     """Testing ..recurrence.next_date_of_same_weekday()."""
 
-    def callFUT(self, wd_src, base_date):
+    def callFUT(self, wd_src, base_date, additional_weeks=0):
         from ..recurrence import next_date_of_same_weekday
-        return next_date_of_same_weekday(wd_src, base_date)
+        return next_date_of_same_weekday(wd_src, base_date, additional_weeks)
 
     def test_weekday_of_wd_src_smaller_than_weekday_of_base_date(self):
         self.assertEqual(self.get_datetime((2014, 7, 28, 10)),
@@ -92,3 +162,8 @@ class NextDateOfSameWeekdayTests(unittest.TestCase,
     def test_wd_src_equal_to_base_date(self):
         dt = self.get_datetime((2014, 7, 23, 17))
         self.assertEqual(dt, self.callFUT(dt, dt))
+
+    def test_additional_whole_weeks_can_be_added(self):
+        self.assertEqual(self.get_datetime((2014, 9, 9, 15)),
+                         self.callFUT(self.get_datetime((2014, 9, 2, 15)),
+                                      self.get_datetime((2014, 9, 1, 15)), 1))